Feat:move page by hash
