+++ css/header.css
@@ -1,0 +1,20 @@
+/* header */
+header{
+  background-color: rgb(255, 255, 255);
+  font-weight: 900;
+  position:fixed;
+  top:0;
+  left:0;
+  right:0;
+  margin-bottom:10px;
+  padding:2rem;
+  box-shadow: 5px 5px 15px lightgray;
+}
+.menu-container{
+  display: inline;
+  position:absolute;
+  right:2rem;
+}
+.menu-container a{
+  padding:0px 20px;
+}
