+++ css/moreInfo.css
@@ -1,0 +1,41 @@
+.detail-container{
+  margin-top:8rem;
+}
+.detail-index{
+  font-size: 2rem;
+  margin-left: 2rem;
+  margin-top: 2rem;
+  font-weight: 900;
+}
+.detail-profile-card{
+  background-color: rgb(30,30,30);
+  color:rgb(173, 255, 80);
+  display: flex;
+  padding:50px 40px;
+  margin:10px 20px;
+  box-shadow: 10px 10px 20px lightgray;
+  border-radius: 15px;
+  align-items:center;
+}
+.detail-profile-image{
+  width:15rem;
+  height:15rem;
+  margin:10px;
+  object-fit: cover;
+}
+.detail-text-container{
+  display:flex;
+  flex-direction: column;
+  margin-left:2rem;
+  color:aliceblue;
+}
+.detail-text-container h1{
+  font-size: 1.5rem;
+  margin-bottom: 1rem;
+  color:rgb(173, 255, 80);
+}
+.detail-text-container .subtitle{
+  margin-top:1rem;
+  font-weight: 900;
+  color:rgb(173, 255, 80);
+}
