+++ css/reset.css
@@ -10,7 +10,3 @@ a {
 img{
   border-radius: 10px;
 }
-h1{
-  margin:4rem 0px;
-  text-align: center;
-}
