+++ index.html
@@ -5,6 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Document</title>
   <link rel="stylesheet" href="./css/reset.css">
+  <link rel="stylesheet" href="./css/header.css">
   <link rel="stylesheet" href="./css/modal.css">
   <link rel="stylesheet" href="./css/main.css">
 </head>
@@ -23,11 +24,6 @@
       </a>
     </div>
   </header>
-  <nav>
-    <ul>
-      <li><a href="./pages/moreInfo/moreInfo.html#div2">2번박스로이동</a></li>
-    </ul>
-  </nav>
   <div id="content"></div>
   <section>
     <h1>Human Resources</h1>
