+++ index.js
@@ -6,6 +6,7 @@ import { getFirestore, collection, addDoc, doc, getDoc, getDocs, orderBy, query,
 
 import {showPreview} from "./js/showPreview.js"
 import * as addModal from "./js/addModal.js";
+import * as addProfileCard from "./js/addProfileCard.js";
 //import {fetchProfileIds} from "./js/fetchProfileIds.js";
 
 export {imgFileInput} ;
@@ -65,45 +66,4 @@ document.querySelector('.btn__upload').addEventListener('click', async function(
       console.log('텍스트를 입력해주세요!');
     }
   }
-})
-
-//profile카드 추가하기 관련 선언
-const profileContainer = document.querySelector('.container');
-let template = `
-    <div>
-      <img class="profile-image" src="{{__profile_image__}}" alt="profile image">
-      <div class="text-container">
-        <h1>{{__profile_name__}}</h1>
-        <span class="subtitle">🌞 POSITION</span>
-        <span>{{__profile_position__}}</span>
-        <span class="subtitle">🔥 GITHUB</span>
-        <span>{{__profile_github__}}</span>
-        <span class="subtitle">📧 EMAIL</span>
-        <span>{{__profile_email__}}</span>
-      </div>
-    </div>
-`;
-//profile카드 추가하기
-const q = query(collection(db,"profiles"))
-let itemNumber = 0;
-onSnapshot(q,(querySnapshot) => {
-  querySnapshot.forEach((doc) => {
-    //console.log('시작:',doc.data())
-    const newProfile = document.createElement("a")
-    newProfile.classList.add('item');
-    //newProfile.setAttribute("href", `https://www.naver.com-${itemNumber}`); //
-    template = template.replace('{{__profile_image__}}',doc.data().image)
-                        .replace('{{__profile_name__}}', doc.data().name)
-                        .replace('{{__profile_position__}}',doc.data().position)
-                        .replace('{{__profile_github__}}',doc.data().github)
-                        .replace('{{__profile_email__}}',doc.data().email);
-    newProfile.innerHTML=template;
-    profileContainer.append(newProfile);
-    template = template.replace(doc.data().image,'{{__profile_image__}}')
-                        .replace( doc.data().name, '{{__profile_name__}}')
-                        .replace(doc.data().position,'{{__profile_position__}}')
-                        .replace(doc.data().github,'{{__profile_github__}}')
-                        .replace(doc.data().email,'{{__profile_email__}}');
-    itemNumber++;
-  });
 })
