+++ pages/moreInfo/moreInfo.html
@@ -5,7 +5,8 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>More Info</title>
   <link rel="stylesheet" href="../../css/reset.css">
-  <link rel="stylesheet" href="../../css/main.css">
+  <link rel="stylesheet" href="../../css/header.css">
+  <link rel="stylesheet" href="../../css/moreInfo.css">
 </head>
 <body>
   <header>
@@ -22,15 +23,7 @@
       </a>
     </div>
   </header>
-  
-  <h2>안녕 여긴 인포페이지</h2>
-
-  <div id="div1" style="background-color: red; height:600px;">
-    1번박스
-  </div>
-  <div id="div2" style="background-color: blue; height:600px;">
-    2번박스
-  </div>
-  <script src="../../js/handleHash.js"></script>
+  <div class="detail-container"></div>
+  <script type="module" src="./moreInfo.js"></script>
 </body>
 </html>
